Fixed the Type values returned by the typed operator implementations in ExecutionEngine/Compiled/Operations. Added support to Compiler.TranslateExpression for unary expressions. Added unary expression type ExecutionEngine/Compiled/Operations/Negation.cs.

+++ QBX/ExecutionEngine/Compiled/Operations/Modulo.cs
@@ -55,7 +55,7 @@ public class LongModulo(IEvaluable left, IEvaluable right) : IEvaluable
 	public CodeModel.Statements.Statement? SourceStatement { get; set; }
 	public CodeModel.Expressions.Expression? SourceExpression { get; set; }
 
-	public DataType Type => DataType.Integer;
+	public DataType Type => DataType.Long;
 
 	public Variable Evaluate(ExecutionContext context)
 	{
